--- a/Cargo.toml
+++ b/Cargo.toml
@@ -227,9 +227,6 @@ index_list = "0.2.11"
 indexmap = "2.2.5"
 indicatif = "0.17.8"
 itertools = "0.10.5"
-jemallocator = { package = "tikv-jemallocator", version = "0.4.1", features = [
-    "unprefixed_malloc_on_supported_platforms",
-] }
 js-sys = "0.3.69"
 json5 = "0.4.1"
 jsonrpc-core = "18.0.0"