--- a/ledger-tool/Cargo.toml
+++ b/ledger-tool/Cargo.toml
@@ -20,7 +20,6 @@ futures = { workspace = true }
 histogram = { workspace = true }
 itertools = { workspace = true }
 log = { workspace = true }
-mimalloc = { workspace = true }
 num_cpus = { workspace = true }
 rayon = { workspace = true }
 regex = { workspace = true }
@@ -56,7 +55,7 @@ thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
-jemallocator = { workspace = true }
+mimalloc = { workspace = true }
 
 [dev-dependencies]
 assert_cmd = { workspace = true }