--- a/ledger-tool/src/main.rs
+++ b/ledger-tool/src/main.rs
@@ -535,11 +535,11 @@ fn assert_capitalization(bank: &Bank) {
 }
 
 #[cfg(not(target_env = "msvc"))]
-use jemallocator::Jemalloc;
+use mimalloc::MiMalloc;
 
 #[cfg(not(target_env = "msvc"))]
 #[global_allocator]
-static GLOBAL: Jemalloc = Jemalloc;
+static GLOBAL: MiMalloc = MiMalloc;
 
 #[allow(clippy::cognitive_complexity)]
 fn main() {