--- a/validator/Cargo.toml
+++ b/validator/Cargo.toml
@@ -27,7 +27,6 @@ jsonrpc-ipc-server = { workspace = true }
 lazy_static = { workspace = true }
 libloading = { workspace = true }
 log = { workspace = true }
-mimalloc = { workspace = true }
 num_cpus = { workspace = true }
 rand = { workspace = true }
 rayon = { workspace = true }
@@ -76,7 +75,7 @@ solana-runtime = { workspace = true, features = ["dev-context-only-utils"] }
 spl-token-2022 = { workspace = true, features = ["no-entrypoint"] }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
-jemallocator = { workspace = true }
+mimalloc = { workspace = true }
 
 [target."cfg(unix)".dependencies]
 libc = { workspace = true }