@ryoqun 2024-05-10
@Lichtso I seems you tried mimalloc before, right? did you find any potential blocker? as far as i can tell, it's better than jemalloc.