@ryoqun 2024-05-10
@alessandrod Note that I'm not opposed to reducing allocs. Obviously, it's better for perf.

I'm just getting tired of jemalloc. While it's well faster than the glibc allocator with less fragmentation, it has its own peculiar perf profile. So, I'm considering replacing it as a separate effort with bp-ing this to v1.17 & v1.18 in mind, parallel to the on-going alloc reduction efforts. I think you're opinionated around mem allocators in general. Is there any concern for using mimalloc? If any, I'd like to address with any additional investigation.

As for reliability (read: no leak & no fragmentation), I'm now running a devbox against mb, needs another day or two, before conclusion.

As a starter, I confirmed this fixes https://github.com/solana-labs/solana/issues/27275:

```
before:
# --block-verification-method=blockstore-processor
ledger processed in 31 seconds, 473 ms
ledger processed in 31 seconds, 164 ms
ledger processed in 31 seconds, 335 ms

# --block-verification-method=unified-scheduler
ledger processed in 16 seconds, 363 ms
ledger processed in 16 seconds, 601 ms
ledger processed in 16 seconds, 512 ms

after:
# --block-verification-method=blockstore-processor
ledger processed in 24 seconds, 665 ms
ledger processed in 24 seconds, 848 ms
ledger processed in 24 seconds, 623 ms

# --block-verification-method=unified-scheduler
ledger processed in 12 seconds, 165 ms
ledger processed in 12 seconds, 78 ms
ledger processed in 12 seconds, 117 ms
```

blockstore-processor is ~27% faster and unified scheduler is ~36% faster according to https://github.com/solana-labs/solana/pull/35286#user-content-steps. I think this alone is quite good perf gain by its own. :)

Also, I haven't confirmed but I'm moderately certain that this fixes the replay stall issue under system overload like rpmalloc fixed the stall (note that rpmalloc is out-of-question due to https://github.com/anza-xyz/agave/pull/1250#issuecomment-2103657904)