@Lichtso 2024-05-10
Yes I tried mimalloc a few months ago and no, nothing problematic stood out to me at the time.