@alessandrod 2024-05-10
> @alessandrod Firstly, note that I'm not opposed to reducing allocs. Obviously, it's better for perf.
> 
> I'm just getting tired of jemalloc. While it's well faster than the glibc allocator with less fragmentation, it has its own peculiar perf profile. So, I'm considering replacing it as a separate effort with bp-ing this to v1.17 & v1.18 in mind, parallel to the on-going alloc reduction efforts. I think you're opinionated around mem allocators in general. Is there any concern for using mimalloc? If any, I'd like to address with any additional investigation.

Uh? Having spent a significant chunk of my time looking at profiles and reading the jemalloc source code my opinion is that it 😂 I'm more than open to replacing jemalloc, I just don't think that we should do it lightly. Changing allocator impacts every subsystem in the validator, personally I think it's a pretty scary change.

> As for reliability (read: no leak & no unbounded fragmentation), I'm now running a devbox against mb, needs another day or two, before conclusion.

Hum imo you can't draw conclusions after running a devbox for a couple of days. Jemalloc sucks, and it's really struggling with the amount and pattern of allocations we do, especially since we do them from so many threads. But at least it isn't crashing 😅 Imagine we switch allocator, and then we find that in some configurations after a while it goes OOM - I don't think it's something we can establish in a couple of days based on a single machine configuration, with a single workload, etc. IMO it requires a lot of testing, especially on nodes with a lot of stake, which are the ones struggling the most with memory.

> As a starter, I confirmed this fixes [solana-labs#27275](https://github.com/solana-labs/solana/issues/27275):
> 
> ```
> before:
> # --block-verification-method=blockstore-processor
> ledger processed in 31 seconds, 473 ms
> ledger processed in 31 seconds, 164 ms
> ledger processed in 31 seconds, 335 ms
> 
> # --block-verification-method=unified-scheduler
> ledger processed in 16 seconds, 363 ms
> ledger processed in 16 seconds, 601 ms
> ledger processed in 16 seconds, 512 ms
> 
> after:
> # --block-verification-method=blockstore-processor
> ledger processed in 24 seconds, 665 ms
> ledger processed in 24 seconds, 848 ms
> ledger processed in 24 seconds, 623 ms
> 
> # --block-verification-method=unified-scheduler
> ledger processed in 12 seconds, 165 ms
> ledger processed in 12 seconds, 78 ms
> ledger processed in 12 seconds, 117 ms
> ```
> 
> blockstore-processor is ~27% faster and unified scheduler is ~36% faster according to [solana-labs#35286 (comment)](https://github.com/solana-labs/solana/pull/35286#user-content-steps). I think this alone is quite good perf gain by its own. :)

This is amazing! 


> Also, I haven't confirmed yet but I'm moderately certain that this fixes the replay stall issue under system overload like rpmalloc fixed the stall (note that rpmalloc is out-of-question due to [#1250 (comment)](https://github.com/anza-xyz/agave/pull/1250#issuecomment-2103657904))

There are at least two stalls that I'm aware of, here I'm assuming you're talking about the one caused by jemalloc's time based decay. Mimalloc almost certainly gets rid of that one, but how? How does it compact/release memory? When? And from what threads? Because getting rid of that stall is also possible by configuring jemalloc. Not saying that we should do it mind, and based on the replay speedups it looks like mimalloc might have better defaults for our workload, I just think that we should understand how mimalloc works and why it's better for us before we switch to it.