@alessandrod 2024-05-10
>  Mimalloc almost certainly gets rid of that one, but how?

It looks like mimalloc does time based purge like jemalloc