@alessandrod 2024-05-10
> It looks like mimalloc does time based purge like jemalloc

I think I'm pretty mimalloc-pilled. I think that the key is that even if it does time based purge like jemalloc, since it has sharded freelists and works at a "mimalloc page" level, it never does gigantic (and contended) deallocations like jemalloc does, which means than even when purging after spikes, deallocation costs are amortized over a bunch of call sites and bounded in time, instead of having seconds long spikes like we do with jemalloc.

I personally would ship this to master and let it ride the trains, and see what happens. While it's in master it'll get tested on the master canaries, and we could ask Joe to test it on at least one big-stake node.