@ryoqun 2024-05-11
thanks for various thoughts and quick look into mimalloc so far! seems you have good first impression of it. lol

I'll spend some time to dig in further and write my own thoughts later.